fixed floating point bug in ia3 update

+++ tests/end2end/model.py
@@ -76,7 +76,7 @@ def low_rank_update(t, rank):
 
 
 def make_ia3_update(value):
-    ia3 = np.random.randn(*value.shape)
+    ia3 = np.random.randn(*value.shape).astype(np.float32)
     axes = (0, -1) if ia3.ndim > 3 else (-1,)
     ia3 = np.mean(ia3, axis=axes, keepdims=True)
     return {"ia3": ia3}
